+++ packages/expo-barcode-scanner/android/src/main/java/expo/modules/barcodescanner/BarCodeScannerModule.java
@@ -7,6 +7,9 @@ import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import expo.modules.barcodescanner.utils.BarCodeScannerResultSerializer;
+import expo.modules.interfaces.barcodescanner.BarCodeScannerInterface;
+import expo.modules.interfaces.barcodescanner.BarCodeScannerResult;
+import expo.modules.interfaces.barcodescanner.BarCodeScannerSettings;
 
 import com.google.android.gms.vision.barcode.Barcode;
 
@@ -20,9 +23,6 @@ import org.unimodules.core.ExportedModule;
 import org.unimodules.core.ModuleRegistry;
 import org.unimodules.core.Promise;
 import org.unimodules.core.interfaces.ExpoMethod;
-import org.unimodules.interfaces.barcodescanner.BarCodeScanner;
-import org.unimodules.interfaces.barcodescanner.BarCodeScannerResult;
-import org.unimodules.interfaces.barcodescanner.BarCodeScannerSettings;
 import org.unimodules.interfaces.imageloader.ImageLoader;
 import org.unimodules.interfaces.permissions.Permissions;
 
@@ -115,7 +115,7 @@ public class BarCodeScannerModule extends ExportedModule {
     imageLoader.loadImageForManipulationFromURL(url, new ImageLoader.ResultListener() {
       @Override
       public void onSuccess(@NonNull Bitmap bitmap) {
-        BarCodeScanner scanner = mBarCodeScannerProvider.createBarCodeDetectorWithContext(getContext());
+        BarCodeScannerInterface scanner = mBarCodeScannerProvider.createBarCodeDetectorWithContext(getContext());
         scanner.setSettings(new BarCodeScannerSettings() {{
           putTypes(types);
         }});