+++ packages/expo-font/android/src/main/java/expo/modules/font/FontLoaderModule.java
@@ -12,8 +12,9 @@ import org.unimodules.core.ExportedModule;
 import org.unimodules.core.ModuleRegistry;
 import org.unimodules.core.Promise;
 import org.unimodules.core.interfaces.ExpoMethod;
-import org.unimodules.interfaces.constants.ConstantsInterface;
 import org.unimodules.interfaces.font.FontManager;
+
+import expo.modules.interfaces.constants.ConstantsInterface;
 
 public class FontLoaderModule extends ExportedModule {
   private static final String ASSET_SCHEME = "asset://";