+++ packages/expo-modules-core/android/src/main/java/expo/modules/interfaces/sensors/services/RotationVectorSensorService.java
@@ -1,6 +1,0 @@
-package expo.modules.interfaces.sensors.services;
-
-import expo.modules.interfaces.sensors.SensorService;
-
-public interface RotationVectorSensorService extends SensorService {
-}