+++ packages/expo-video-thumbnails/ios/EXVideoThumbnails/EXVideoThumbnailsModule.m
@@ -16,7 +16,7 @@ UM_EXPORT_MODULE(ExpoVideoThumbnails);
 - (void)setModuleRegistry:(UMModuleRegistry *)moduleRegistry
 {
   _moduleRegistry = moduleRegistry;
-  _fileSystem = [moduleRegistry getModuleImplementingProtocol:@protocol(UMFileSystemInterface)];
+  _fileSystem = [moduleRegistry getModuleImplementingProtocol:@protocol(EXFileSystemInterface)];
 }
 
 UM_EXPORT_METHOD_AS(getThumbnail,
@@ -30,7 +30,7 @@ UM_EXPORT_METHOD_AS(getThumbnail,
     if (!_fileSystem) {
       return reject(@"E_MISSING_MODULE", @"No FileSystem module.", nil);
     }
-    if (!([_fileSystem permissionsForURI:url] & UMFileSystemPermissionRead)) {
+    if (!([_fileSystem permissionsForURI:url] & EXFileSystemPermissionRead)) {
       return reject(@"E_FILESYSTEM_PERMISSIONS", [NSString stringWithFormat:@"File '%@' isn't readable.", source], nil);
     }
   }