+++ packages/expo-av/CHANGELOG.md
@@ -15,6 +15,10 @@
 - Fixed the web Video Fullscreen APIs in Safari ([#12258](https://github.com/expo/expo/pull/12258) by [@elliotdickison](https://github.com/elliotdickison))
 - Enable kotlin in all modules. ([#12716](https://github.com/expo/expo/pull/12716) by [@wschurman](https://github.com/wschurman))
 
+### 💡 Others
+
+- Migrated from `unimodules-file-system-interface` to `expo-modules-core`.
+
 ## 9.1.2 — 2021-04-13
 
 _This version does not introduce any user-facing changes._